backend/requirements.txt
fastapi==0.104.1
sqlmodel==0.0.8
pydantic==1.10.13
pydantic-settings==2.1.0
uvicorn==0.24.0
python-multipart==0.0.6
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.41
psycopg2-binary==2.9.9
httpx==0.25.2
python-dotenv==1.0.0